% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brite_ids.R
\name{brite_ids}
\alias{brite_ids}
\title{Functions returns kegg brite ids}
\usage{
brite_ids()
}
\arguments{
\item{()}{no need to add something here, just ()}
}
\value{
kegg brite ids
}
\description{
Just type "brite_ids()" to get brite ids and names of all hierachical files in kegg
}
