% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check is a prelim function for Kostas}
\usage{
check(x)
}
\arguments{
\item{ids}{list of ko genes}
}
\value{
list with kegg genes and respective kos
}
\description{
This function takes as input a list of ko ids in the format of kegg genes and returns a list!
}
\details{
Example: ids <- as.list(c("hsa:5053","cfr:102516611","lve:103079765"))
out <- as.data.frame(do.call(rbind,lapply(ids, check)))
}
