% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ko_idToBrite.R
\name{Ko_idToBrite}
\alias{Ko_idToBrite}
\title{Ko_idToBrite: retrieve brite inforamtion for ko ids}
\usage{
Ko_idToBrite(id)
}
\arguments{
\item{id}{vector of ko ids in the format ko:KXXXXX}
}
\value{
dataframe with ko ids and associated kegg brite information
}
\description{
This function takes as input a vector of ko ids in the format of ko:KXXXXX and returns a data frame
with associated kegg brite inforamtions. The brite information do not include the pathway mappings.
Pathway mappings can be retrieved with the function MapKo_idToPathway
The function takes a little bit to run, it needs to download and format different datasets
}
\details{
Example: id <- id <- c("ko:K00031","ko:K00001","ko:K00002")
Ko_idToBrite(id=id)
}
