% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeKeggReferenceTableFromBrite.R
\name{MakeKeggReferenceTableFromBrite}
\alias{MakeKeggReferenceTableFromBrite}
\title{MakeKeggReferenceTable: creates a reference table for your Kegg hierachical files}
\usage{
MakeKeggReferenceTableFromBrite(brite)
}
\arguments{
\item{brite}{Kegg brite ID}
}
\value{
A data frame with kegg hierachy files
}
\description{
This function takes as input a brite ID and convert the hierachical files into a dataframe.
The Kegg brite IDs can be retrieved with the helper function brite_ids()
Example: MakeKeggReferenceTableFromBrite(brite="br:br08901") returns the pathway calssifications
}
