% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapKo_idToPathway.R
\name{MapKo_idToPathway}
\alias{MapKo_idToPathway}
\title{MapKo_idToPathway: retrieve pathway inforamtion for ko ids}
\usage{
MapKo_idToPathway(id)
}
\arguments{
\item{id}{vector of ko ids in the format ko:KXXXXX}
}
\value{
dataframe with ko ids and associated kegg pathway information
}
\description{
This function takes as input a vector of ko ids in the format of ko:KXXXXX and returns a data frame
with associated kegg pathway inforamtions. If you also want to have the respective module information,
run afterwards the function 'ModulesToPath'
}
\details{
Example: id <- id <- c("ko:K00031","ko:K00001","ko:K00002")
MapKo_idToPathway(id=id)
}
