% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulesToPath.R
\name{ModulesToPath}
\alias{ModulesToPath}
\title{ModulesToPath: adds module inforamtion to the output of MapKo_idToPathway}
\usage{
ModulesToPath(path)
}
\arguments{
\item{path}{output data frame of the MapKo_idToPathway}
}
\value{
dataframe with module inforamtion added to the output of MapKo_idToPathway
}
\description{
Example: id <- c("ko:K00031","ko:K00001","ko:K00002")
out <- MapKo_idToPathway(id=id)
ModulesToPath(out)
}
