% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot_annotations.R
\name{uniprot_annotations}
\alias{uniprot_annotations}
\title{Retrieve IDs from the uniprot REST API}
\usage{
uniprot_annotations(ids, database)
}
\arguments{
\item{ids}{vector of uniprot IDs}

\item{database}{vector of database abbreviations}
}
\value{
A data frame with all requested IDs
}
\description{
This functions retrives IDs from the uniprot server. It takes as input swiss
prot identifers that generated in the trinnoate output table, or that are
retrieved form a blast search against the swiss prot database (it also works
with Trembl IDs)
You have to provide a vector ID names e.g ID = c("MSRA_PSEPW","NAT6_HUMAN")
and specifiy your mapping options. A complete list of mapping options is
available at https://www.uniprot.org/help/api_idmapping
Examples of mapping options (database options):
database=c('GENENAME','ACC','EGGNOG_ID','KEGG_ID')
This function connects to the uniprot REST API and returns the specified
options in the form of a data frame.
}
